+++ src/main/java/subway/line/application/LineDto.java
@@ -1,21 +1,30 @@
-package subway.line;
+package subway.line.application;
 
-public class LineRequest {
+import subway.line.presentation.LineRequest;
+
+public class LineDto {
     private String name;
     private String color;
     private Long upStationId;
     private Long downStationId;
     private Long distance;
 
-    public LineRequest() {
-    }
-
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
+    private LineDto(String name, String color, Long upStationId, Long downStationId, Long distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
+    }
+
+    public static LineDto from(LineRequest lineRequest) {
+        return new LineDto(
+                lineRequest.getName(),
+                lineRequest.getColor(),
+                lineRequest.getUpStationId(),
+                lineRequest.getDownStationId(),
+                lineRequest.getDistance()
+        );
     }
 
     public String getName() {
