+++ src/main/java/subway/line/application/LineService.java
@@ -1,8 +1,11 @@
-package subway.line;
+package subway.line.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.station.StationQuery;
+import subway.line.domain.Line;
+import subway.line.domain.LineRepository;
+import subway.line.presentation.LineResponse;
+import subway.station.application.StationQuery;
 
 import java.util.List;
 import java.util.stream.Collectors;
