+++ src/main/java/subway/line/domain/LineRepository.java
@@ -1,6 +1,7 @@
-package subway.line;
+package subway.line.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.line.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
 }
