+++ src/main/java/subway/line/presentation/LineRequest.java
@@ -1,28 +1,21 @@
-package subway.line;
+package subway.line.presentation;
 
-public class LineDto {
+public class LineRequest {
     private String name;
     private String color;
     private Long upStationId;
     private Long downStationId;
     private Long distance;
 
-    private LineDto(String name, String color, Long upStationId, Long downStationId, Long distance) {
+    public LineRequest() {
+    }
+
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
-    }
-
-    public static LineDto from(LineRequest lineRequest) {
-        return new LineDto(
-                lineRequest.getName(),
-                lineRequest.getColor(),
-                lineRequest.getUpStationId(),
-                lineRequest.getDownStationId(),
-                lineRequest.getDistance()
-        );
     }
 
     public String getName() {
