+++ src/main/java/subway/line/presentation/LineResponse.java
@@ -1,6 +1,7 @@
-package subway.line;
+package subway.line.presentation;
 
-import subway.station.StationResponse;
+import subway.line.domain.Line;
+import subway.station.presentation.StationResponse;
 
 import java.util.List;
