+++ src/main/java/subway/line/presentation/UpdateLineRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.presentation;
 
 import subway.common.Request;
