+++ src/main/java/subway/station/application/StationService.java
@@ -1,7 +1,12 @@
-package subway.station;
+package subway.station.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.station.application.StationQuery;
+import subway.station.domain.Station;
+import subway.station.domain.StationRepository;
+import subway.station.presentation.StationRequest;
+import subway.station.presentation.StationResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
