+++ src/main/java/subway/station/domain/StationRepository.java
@@ -1,6 +1,7 @@
-package subway.station;
+package subway.station.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.station.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
