+++ src/main/java/subway/station/presentation/StationController.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import subway.station.application.StationService;
 
 import java.net.URI;
 import java.util.List;
