+++ src/test/java/subway/station/StationFixtures.java
@@ -1,7 +1,9 @@
 package subway.station;
 
+import subway.station.presentation.StationRequest;
+
 public class StationFixtures {
-    public static final StationRequest  강남역_생성_요청= new StationRequest("강남역");
+    public static final StationRequest 강남역_생성_요청= new StationRequest("강남역");
     public static final StationRequest 서울대입구역_생성_요청= new StationRequest("서울대입구역");
     public static final StationRequest 신논현역_생성_요청= new StationRequest("신논현역");
     public static final StationRequest 낙성대역_생성_요청= new StationRequest("낙성대역");
